pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# --- PROJECT METADATA & DEPENDENCIES ---
[project]
name = "ml_core"
version = "0.1.0"
description = "MLLM for Barrett's Oesophagus AMC"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core Data & Science
    "h5py",
    "humanize",
    "numpy",
    "pandas",
    "scipy",
    "scikit-learn",
    "typing_extensions",

    # Machine Learning & Deep Learning (Torch/HF)
    "accelerate",
    "bitsandbytes",
    "huggingface-hub",
    "safetensors",
    "tensorboard",
    "torch",
    "torcheval",
    "torchvision",
    "transformers",
    "umap-learn",

    # Image Processing
    "imageio",
    "scikit-image",

    # Visualization
    "matplotlib",
    "rich",
    "seaborn",

    # Utilities
    "ninja",
    "python-dotenv",
    "pyyaml",
    "tqdm",

    # Dev Tools
    "pre-commit",
    "pytest",
    "ruff",
]

# --- TOOL CONFIGURATION ---
[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [".git", "__pycache__", "venv"]
# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
# E: pycodestyle errors
# F: Pyflakes
# I: isort (Import Sorting - MLOps critical!)
# B: flake8-bugbear (Likely bugs)
select = ["E", "F", "I", "B"]
# B905: no strict=True needed for zip
ignore = ["E501", "B905", "E741"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
